✨ feat: 직원관리

+++ src/main/java/org/ateam/oncare/employee/command/dto/EmployeeRequestDTO.java
@@ -13,29 +13,29 @@ import java.util.List;
 @AllArgsConstructor
 @NoArgsConstructor
 public class EmployeeRequestDTO {
-// --- 기본 정보 ---
-private String name;
-private String address; // 오타 수정됨 (adress -> address)
-private String phone;
-private String email;
-private LocalDate birth;
-private String gender;
-private String emergencyNumber;
-private LocalDate hireDate;
-private LocalDate endDate;
+    // --- 기본 정보 ---
+    private String name;
+    private String address; // 오타 수정됨 (adress -> address)
+    private String phone;
+    private String email;
+    private LocalDate birth;
+    private String gender;
+    private String emergencyNumber;
+    private LocalDate hireDate;
+    private LocalDate endDate;
 
-// --- [중요] 연관 테이블의 ID(PK) 값만 받습니다 ---
-// 이름(예: "간호팀")을 받는 게 아니라 ID(예: 1)를 받습니다.
-private Long deptCode;  // m_department 테이블의 id
-private Long jobCode;   // m_job 테이블의 id
-private Long statusId;  // m_status 테이블의 status (PK)
-private Long managerId; // 매니저 직원 ID
+    // --- [중요] 연관 테이블의 ID(PK) 값만 받습니다 ---
+    // 이름(예: "간호팀")을 받는 게 아니라 ID(예: 1)를 받습니다.
+    private Integer deptCode; // m_department 테이블의 id
+    private Long jobCode; // m_job 테이블의 id
+    private Long statusId; // m_status 테이블의 status (PK)
+    private Long managerId; // 매니저 직원 ID
 
-// 경력 리스트
-private List<CareerDTO> careers;
+    // 경력 리스트
+    private List<CareerDTO> careers;
 
-// 보유 자격증 리스트
-private List<CertificateInputDTO> certificates;
+    // 보유 자격증 리스트
+    private List<CertificateInputDTO> certificates;
 
     // 서비스 유형 ID 리스트 (예: [1, 2])
     private List<Long> serviceTypeIds;
