+++ src/main/java/org/ateam/oncare/employee/query/mapper/EmployeeMapper.java
@@ -27,7 +27,7 @@ public interface EmployeeMapper {
     List<EducationDTO> selectEducations(Long employeeId);
 
     // 요양보호사(담당수급자) 일정 조회
-    List<VisitScheduleVO> selectSchedulesByEmployeeId(Long employeeId);
+    List<VisitScheduleVO> selectSchedulesByEmployeeId(Integer employeeId);
 
     // 방문 일정 조회 (상세)
     List<VisitScheduleVO> selectVisitSchedules(java.util.Map<String, Object> params);
